UX improvements to submit button and action links
Removed the float:right for the form submit buttons, keeping the button and the action links on their own lines rather than side by side, which should help avoid confusion about what to click on. 

Also removed the negative margin from the top of the button, as this was cutting off the top of that button on the reset password form.