Prefix all template classes
Since it's inception, TML was intended to be an exact copy of
wp-login.php for the front end. This included class names. However, it
is now apparent that generic class names that are used for that one
specific page, such as 'login', are not suitable for a front end page.
Many theme stylesheets apply their own styling to these generic classes.
This can potentially break a site's layout when using TML's stylesheet.
So, all classes are now prefixed with 'tml-'.

Fixes: #6.