Don't redirect to TML pages when using referer for redirect type
When you start a password reset, you are sent an email with a link to
reset your password. When you click this link, there will be no original
referer. After you submit the password reset form, you're redirected to
the login form with a confirmation message. However, the referer now
becomes the password reset page because there is no original referer.
This, of course, causes an "invalid key" error, thus making the user
think that they did not actually reset their password. This fix will
simply use the default URL if it detects that the redirect URL is a TML
page.

Fixes: #27