Cookie Redirection
I'm working on a site where we want the user to return to the last page they visited after logging in. 
To do this I set a cookie in any pages a user might visit before logging in. 
I added an option in the admin under the login option that you can set a cookie name use as a redirection. Please check the custom redirection admin file the switch case. If no cookie by that name is set or cookies are not on, use default.